# app/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Flask App",
    "description": "APIs for user management",
    "version": "1.0"
  },
  "paths": {
    "/users": {
      "get": {
        "summary": "Get all users",
        "responses": {
          "200": {
            "description": "Successful response",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/User" }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new user",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/User" }
          }
        ],
        "responses": {
          "201": { "description": "User created" },
          "400": { "description": "Invalid input" }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "summary": "Get a user by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "string" }
        ],
        "responses": {
          "200": {
            "description": "User details",
            "schema": { "$ref": "#/definitions/User" }
          },
          "404": { "description": "User not found" }
        }
      },
      "put": {
        "summary": "Update a user",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "string" },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/UserUpdate" }
          }
        ],
        "responses": {
          "200": { "description": "User updated" },
          "404": { "description": "User not found" }
        }
      },
      "delete": {
        "summary": "Delete a user",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "string" }
        ],
        "responses": {
          "200": { "description": "User deleted" },
          "404": { "description": "User not found" }
        }
      }
    },
    "/login": {
      "post": {
        "summary": "User login",
        "parameters": [
          { "name": "email", "in": "query", "required": true, "type": "string" },
          { "name": "password", "in": "query", "required": true, "type": "string" }
        ],
        "responses": {
          "200": {
            "description": "Login successful",
            "schema": {
              "type": "object",
              "properties": {
                "token": { "type": "string" }
              }
            }
          },
          "401": { "description": "Invalid credentials" }
        }
      }
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "required": ["name", "email", "password"],
      "properties": {
        "name": { "type": "string" },
        "email": { "type": "string" },
        "password": { "type": "string" }
      }
    },
    "UserUpdate": {
      "type": "object",
      "properties": {
        "name": { "type": "string" },
        "email": { "type": "string" },
        "password": { "type": "string" }
      }
    }
  }
}